extract code from cluster privilege to cluster privilege resolver
This commit extracts the code related to converting name to `ClusterPrivilege`
from `ClusterPrivilege` class into `ClusterPrivilegeResolver`
 and updates the references.

Relates #44048
