+++ x-pack/plugin/core/src/test/java/org/elasticsearch/xpack/core/security/authz/permission/LimitedRoleTests.java
@@ -21,7 +21,7 @@ import org.elasticsearch.transport.TransportRequest;
 import org.elasticsearch.xpack.core.security.authz.accesscontrol.IndicesAccessControl;
 import org.elasticsearch.xpack.core.security.authz.privilege.ApplicationPrivilege;
 import org.elasticsearch.xpack.core.security.authz.privilege.ApplicationPrivilegeDescriptor;
-import org.elasticsearch.xpack.core.security.authz.privilege.ClusterPrivilege;
+import org.elasticsearch.xpack.core.security.authz.privilege.ClusterPrivilegeResolver;
 import org.elasticsearch.xpack.core.security.authz.privilege.IndexPrivilege;
 import org.junit.Before;
 
@@ -200,27 +200,27 @@ public class LimitedRoleTests extends ESTestCase {
     public void testCheckClusterPrivilege() {
         Role fromRole = Role.builder("a-role").cluster(Collections.singleton("manage_security"), Collections.emptyList())
                 .build();
-        assertThat(fromRole.grants(ClusterPrivilege.ALL), is(false));
-        assertThat(fromRole.grants(ClusterPrivilege.MANAGE_SECURITY), is(true));
+        assertThat(fromRole.grants(ClusterPrivilegeResolver.ALL), is(false));
+        assertThat(fromRole.grants(ClusterPrivilegeResolver.MANAGE_SECURITY), is(true));
 
         {
             Role limitedByRole = Role.builder("scoped-role")
                     .cluster(Collections.singleton("all"), Collections.emptyList()).build();
-            assertThat(limitedByRole.grants(ClusterPrivilege.ALL), is(true));
-            assertThat(limitedByRole.grants(ClusterPrivilege.MANAGE_SECURITY), is(true));
+            assertThat(limitedByRole.grants(ClusterPrivilegeResolver.ALL), is(true));
+            assertThat(limitedByRole.grants(ClusterPrivilegeResolver.MANAGE_SECURITY), is(true));
             Role role = LimitedRole.createLimitedRole(fromRole, limitedByRole);
-            assertThat(role.grants(ClusterPrivilege.ALL), is(false));
-            assertThat(role.grants(ClusterPrivilege.MANAGE_SECURITY), is(true));
+            assertThat(role.grants(ClusterPrivilegeResolver.ALL), is(false));
+            assertThat(role.grants(ClusterPrivilegeResolver.MANAGE_SECURITY), is(true));
         }
         {
             Role limitedByRole = Role.builder("scoped-role")
                     .cluster(Collections.singleton("monitor"), Collections.emptyList()).build();
-            assertThat(limitedByRole.grants(ClusterPrivilege.ALL), is(false));
-            assertThat(limitedByRole.grants(ClusterPrivilege.MONITOR), is(true));
+            assertThat(limitedByRole.grants(ClusterPrivilegeResolver.ALL), is(false));
+            assertThat(limitedByRole.grants(ClusterPrivilegeResolver.MONITOR), is(true));
             Role role = LimitedRole.createLimitedRole(fromRole, limitedByRole);
-            assertThat(role.grants(ClusterPrivilege.ALL), is(false));
-            assertThat(role.grants(ClusterPrivilege.MANAGE_SECURITY), is(false));
-            assertThat(role.grants(ClusterPrivilege.MONITOR), is(false));
+            assertThat(role.grants(ClusterPrivilegeResolver.ALL), is(false));
+            assertThat(role.grants(ClusterPrivilegeResolver.MANAGE_SECURITY), is(false));
+            assertThat(role.grants(ClusterPrivilegeResolver.MONITOR), is(false));
         }
     }
